+++ src/include/s6/s6-supervise.h
@@ -13,8 +13,8 @@
 #define S6_SVSTATUS_SIZE 35
 
 extern int s6_svc_ok (char const *) ;
-extern int s6_svc_write (char const *, char const *, unsigned int) ;
-extern int s6_svc_writectl (char const *, char const *, char const *, unsigned int) ;
+extern int s6_svc_write (char const *, char const *, size_t) ;
+extern int s6_svc_writectl (char const *, char const *, char const *, size_t) ;
 extern int s6_svc_main (int, char const *const *, char const *, char const *, char const *) ;
 
 typedef struct s6_svstatus_s s6_svstatus_t, *s6_svstatus_t_ref ;
