+++ src/libs6/s6_svc_writectl.c
@@ -1,12 +1,13 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <skalibs/bytestr.h>
 #include <s6/s6-supervise.h>
 
-int s6_svc_writectl (char const *service, char const *subdir, char const *s, unsigned int len)
+int s6_svc_writectl (char const *service, char const *subdir, char const *s, size_t len)
 {
-  unsigned int svlen = str_len(service) ;
-  unsigned int sublen = str_len(subdir) ;
+  size_t svlen = str_len(service) ;
+  size_t sublen = str_len(subdir) ;
   char fn[svlen + sublen + 10] ;
   byte_copy(fn, svlen, service) ;
   fn[svlen] = '/' ;
